refactor all hasura webhooks
