+++ src/controllers/api/v1/index.ts
@@ -7,7 +7,10 @@ export * from "./files";
 export * from "./submission";
 export * from "./urlInfo";
 export * from "./version";
-export * from "./webhooks/hasura";
 export * from "./webhooks/kratos";
 export * from "./webhooks/patch";
 export * from "./webhooks/hasura/account";
+export * from "./webhooks/hasura/auth";
+export * from "./webhooks/hasura/chat";
+export * from "./webhooks/hasura/checkworthiness";
+export * from "./webhooks/hasura/trigger";
